bau change thaya chhe bhai

+++ src/main/java/com/project/ims/Services/AdminService.java
@@ -1,60 +1,45 @@
 package com.project.ims.Services;
 
+// imports 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
-
 import com.project.ims.IServices.IAdminService;
 import com.project.ims.Models.Admin;
 import com.project.ims.Repo.AdminRepo;
-import com.project.ims.Repo.CustomerRepo;
-import com.project.ims.Repo.DeliveryManRepo;
-import com.project.ims.Repo.OrderRepo;
-import com.project.ims.Repo.ProductRepo;
-import com.project.ims.Repo.SupplierRepo;
-import com.project.ims.Repo.WManagerRepo;
-import com.project.ims.Repo.WareHouseRepo;
 
 @Component
 @Service
 public class AdminService implements IAdminService {
 
+    // necessary dependency Injections
     @Autowired
     private AdminRepo adminRepo;
 
-    @Autowired
-    private CustomerRepo customerRepo;
 
-    @Autowired
-    private OrderRepo orderRepo;
-
-    @Autowired
-    private DeliveryManRepo deliveryManRepo;
-
-    @Autowired
-    private ProductRepo productRepo;
-
-    @Autowired
-    private SupplierRepo supplierRepo;
-
-    @Autowired
-    private WareHouseRepo wareHouseRepo;
-
-    @Autowired
-    private WManagerRepo wManagerRepo;
+    // Services
 
 
     @Override
     public Admin getAdminById(String id) {
+
+        if (id == null)
+        {
+            throw new RuntimeException("Id shouldn't be null");
+        }
+
         return adminRepo.findById(id).orElse(null);
     }
 
     @Override
     public Admin addAdmin(Admin admin) {
-
-        if(admin.getId() == null || admin.getId().isEmpty())
+        
+        if(admin == null)
+        {
+            throw new RuntimeException("Admin data shouldn't be null");
+        }
+        else if(admin.getId() == null || admin.getId().isEmpty())
         {
             throw new RuntimeException("Admin ID cannot be empty");
         }
@@ -72,18 +57,32 @@ public class AdminService implements IAdminService {
 
     @Override
     public Admin updateAdmin(Admin admin) {
+
+        // checking if admin data is null or not
+        if (admin == null)
+        {
+            throw new RuntimeException("Admin data shouldn't be null");
+        }
+
         return adminRepo.save(admin);
     }
 
     @Override
     public void deleteAdmin(String id) {
 
-        if (adminRepo.existsById(id))
+        // checking if id is valid or not
+        if (id != null && adminRepo.existsById(id))
         {
             adminRepo.deleteById(id);
         }
+        else
+        {
+            throw new RuntimeException("Admin doesn't exist");
+        }
     }
 
+
+    // get all admins
     @Override
     public List<Admin> getAllAdmin() {
         return adminRepo.findAll();
